requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart

# Database (MongoDB)
motor>=3.3.0
pymongo>=4.6.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Authentication
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0

# Rate Limiting
slowapi

# Environment
python-dotenv

# Validation
pydantic>=2.0.0
pydantic[email]

# LLM Providers
google-generativeai>=0.3.0
httpx>=0.25.0

# PDF Processing
pymupdf
Pillow

# File Type Detection
python-magic-bin;sys_platform=='win32'
python-magic;sys_platform!='win32'

# Export
openpyxl

# Metrics & Logging
prometheus-client
